# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Couvoit API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/trips": {
            "get": {
                "tags": [
                    "Trips"
                ],
                "summary": "Search trips",
                "description": "Search trips.\n\nQuery: startingcity, arrivalcity, tripdate, per_page",
                "operationId": "tripsIndex",
                "parameters": [
                    {
                        "name": "startingcity",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "arrivalcity",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tripdate",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Trips"
                ],
                "summary": "Create trip",
                "description": "Create a trip.",
                "operationId": "tripsStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTripRequestPayload"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/trips/{id}": {
            "get": {
                "tags": [
                    "Trips"
                ],
                "summary": "Get trip by id",
                "description": "Show a trip.",
                "operationId": "tripsShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Trips"
                ],
                "summary": "Delete trip permanently",
                "description": "Permanently delete a trip.",
                "operationId": "tripsDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Trips"
                ],
                "summary": "Update trip",
                "description": "Update a trip.",
                "operationId": "tripsUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTripRequestPayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/trips/{id}/person": {
            "get": {
                "tags": [
                    "Trips"
                ],
                "summary": "List trip passengers",
                "description": "List passengers of a trip.",
                "operationId": "tripsPassengers",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Trips"
                ],
                "summary": "Reserve a seat",
                "description": "Reserve a seat for a trip.",
                "operationId": "tripsReserve",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReserveTripRequestPayload"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "200": {
                        "description": "Already exists"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/trips/{id}/cancel": {
            "patch": {
                "tags": [
                    "Trips"
                ],
                "summary": "Cancel trip",
                "description": "Cancel a trip.",
                "operationId": "tripsCancel",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/trips/{id}/reservations": {
            "delete": {
                "tags": [
                    "Trips"
                ],
                "summary": "Cancel reservation",
                "description": "Cancel a reservation for this trip.",
                "operationId": "tripsCancelReservation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CancelReservationRequestPayload"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "ReserveTripRequestPayload": {
                "required": [
                    "person_id"
                ],
                "properties": {
                    "person_id": {
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
            "StoreTripRequestPayload": {
                "required": [
                    "kms",
                    "trip_datetime",
                    "available_seats",
                    "starting_address",
                    "arrival_address"
                ],
                "properties": {
                    "kms": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0.01,
                        "example": 12.5
                    },
                    "trip_datetime": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-02-18T14:00:00+01:00"
                    },
                    "available_seats": {
                        "type": "integer",
                        "maximum": 9,
                        "minimum": 1,
                        "example": 3
                    },
                    "smoking_allowed": {
                        "type": "boolean",
                        "example": false,
                        "nullable": true
                    },
                    "starting_address": {
                        "required": [
                            "street_number",
                            "street_name",
                            "postal_code",
                            "city_name"
                        ],
                        "properties": {
                            "street_number": {
                                "type": "string",
                                "maxLength": 50,
                                "example": "10"
                            },
                            "street_name": {
                                "type": "string",
                                "maxLength": 255,
                                "example": "Rue de Rivoli"
                            },
                            "postal_code": {
                                "type": "string",
                                "maxLength": 20,
                                "example": "75001"
                            },
                            "city_name": {
                                "type": "string",
                                "maxLength": 255,
                                "example": "Paris"
                            }
                        },
                        "type": "object"
                    },
                    "arrival_address": {
                        "required": [
                            "street_number",
                            "street_name",
                            "postal_code",
                            "city_name"
                        ],
                        "properties": {
                            "street_number": {
                                "type": "string",
                                "maxLength": 50,
                                "example": "20"
                            },
                            "street_name": {
                                "type": "string",
                                "maxLength": 255,
                                "example": "Avenue de France"
                            },
                            "postal_code": {
                                "type": "string",
                                "maxLength": 20,
                                "example": "75013"
                            },
                            "city_name": {
                                "type": "string",
                                "maxLength": 255,
                                "example": "Paris"
                            }
                        },
                        "type": "object"
                    },
                    "person_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UpdateTripRequestPayload": {
                "properties": {
                    "kms": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0.01,
                        "example": 15,
                        "nullable": true
                    },
                    "trip_datetime": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-02-18T16:00:00+01:00",
                        "nullable": true
                    },
                    "available_seats": {
                        "type": "integer",
                        "maximum": 9,
                        "minimum": 1,
                        "example": 2,
                        "nullable": true
                    },
                    "smoking_allowed": {
                        "type": "boolean",
                        "example": true,
                        "nullable": true
                    },
                    "starting_address": {
                        "properties": {
                            "street_number": {
                                "type": "string",
                                "maxLength": 50,
                                "example": "10",
                                "nullable": true
                            },
                            "street_name": {
                                "type": "string",
                                "maxLength": 255,
                                "example": "Rue de Rivoli",
                                "nullable": true
                            },
                            "postal_code": {
                                "type": "string",
                                "maxLength": 20,
                                "example": "75001",
                                "nullable": true
                            },
                            "city_name": {
                                "type": "string",
                                "maxLength": 255,
                                "example": "Paris",
                                "nullable": true
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "arrival_address": {
                        "properties": {
                            "street_number": {
                                "type": "string",
                                "maxLength": 50,
                                "example": "20",
                                "nullable": true
                            },
                            "street_name": {
                                "type": "string",
                                "maxLength": 255,
                                "example": "Avenue de France",
                                "nullable": true
                            },
                            "postal_code": {
                                "type": "string",
                                "maxLength": 20,
                                "example": "75013",
                                "nullable": true
                            },
                            "city_name": {
                                "type": "string",
                                "maxLength": 255,
                                "example": "Paris",
                                "nullable": true
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}